fix paths

+++ jenkins/common/setup_docker.sh
@@ -3,13 +3,14 @@ DOCKER_NAME="release-test-automation-${DOCKER_SUFFIX}"
 
 DOCKER_TAG="${DOCKER_NAME}:$(cat containers/this_version.txt)${ARCH}"
 DOCKER_NAMESPACE="arangodb/"
-if ${DOCKER} pull "${REGISTRY_URL}${DOCKER_NAMESPACE}${DOCKER_TAG}"; then
-    echo "using ready built container"
-else
-    ${DOCKER} build "containers/${DOCKER}_$(echo "${DOCKER_SUFFIX}"|sed "s;-;_;g")${ARCH}"  -t "${DOCKER_TAG}" || exit
-    DOCKER_NAMESPACE=""
+if test "${MODE}" != "native"; then
+  if ${DOCKER} pull "${REGISTRY_URL}${DOCKER_NAMESPACE}${DOCKER_TAG}"; then
+      echo "using ready built container"
+  else
+      ${DOCKER} build "containers/${DOCKER}_$(echo "${DOCKER_SUFFIX}"|sed "s;-;_;g")${ARCH}"  -t "${DOCKER_TAG}" || exit
+      DOCKER_NAMESPACE=""
+  fi
 fi
-
 . ./jenkins/common/pre_cleanup_docker.sh
 
 ${DOCKER} network create $DOCKER_NETWORK_NAME
@@ -34,8 +35,13 @@ DOCKER_ARGS=(
          -v /dev/shm:/dev/shm \
         )
 
-TRAP_CLEANUP=(
-    "${DOCKER} kill ${DOCKER_NAME}"
+TRAP_CLEANUP=()
+if test "${MODE}" != "native"; then
+    TRAP_CLEANUP+=(
+        "${DOCKER} kill ${DOCKER_NAME}"
+    )
+fi
+TRAP_CLEANUP+=(
     "${DOCKER} rm ${DOCKER_NAME}"
     "${TRAP_CLEANUP[@]}"
 )
